Split VASP tests to chgcar, locpot, & poscar
